cryptoapp/requirements/base.txt
Django
djangorestframework
django-cors-headers
gunicorn
psycopg2
dj-database-url
dj-static
django-appconf
static3
wsgiref
django-configurations
django-rest-auth
django-allauth
django-phonenumber-field
authy